sessions/session-01-ai-chatbots/starter-template/requirements.txt
google-cloud-aiplatform>=1.26.0
vertexai>=0.1.0
python-dotenv>=1.0.0
